msvs: add ARM64 build file generation
GYP already understands ARM64 from the other platforms it supports, but
it did not properly generate ARM ASM rules for Visual Studio. This
change adds support for it, from including the right props sheet to
emitting MARMASM tags instead of MASM tags. The tags required plumbing
the target architecture through a few layers of abstractions so the
right file types could be associated with it.